refactor(api): split Spotify tests into routes_tests_spotify.rs to honor 1000-line cap
