+++ crates/sp-server/src/api/routes_tests.rs
@@ -873,134 +873,3 @@ async fn ndi_health_endpoint_returns_seeded_pipeline() {
     assert_eq!(arr[0]["ndi_name"].as_str(), Some("SP-test"));
     assert_eq!(arr[0]["state"], serde_json::json!("Playing"));
 }
-
-// ---------------------------------------------------------------------------
-// PATCH /api/v1/videos/{id} — spotify_url field (A.2)
-// ---------------------------------------------------------------------------
-
-async fn insert_test_video(state: &AppState) -> i64 {
-    sqlx::query(
-        "INSERT INTO playlists (name, youtube_url, ndi_output_name, scene_name, is_active) \
-         VALUES ('p', 'https://youtube.com/playlist?list=PLtest', 'n', 's', 1)",
-    )
-    .execute(&state.pool)
-    .await
-    .unwrap();
-    let row: (i64,) = sqlx::query_as(
-        "INSERT INTO videos (playlist_id, youtube_id, title) \
-         VALUES (1, 'aaaaaaaaaaa', 't') RETURNING id",
-    )
-    .fetch_one(&state.pool)
-    .await
-    .unwrap();
-    row.0
-}
-
-#[tokio::test]
-async fn patch_video_extracts_spotify_track_id_from_url() {
-    let state = test_state().await;
-    let video_id = insert_test_video(&state).await;
-    let app = app(state.clone());
-
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/v1/videos/{video_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(
-                    serde_json::to_string(&serde_json::json!({
-                        "spotify_url": "https://open.spotify.com/track/3n3Ppam7vgaVa1iaRUc9Lp?si=ab"
-                    }))
-                    .unwrap(),
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NO_CONTENT);
-
-    let stored = db::models::get_video_spotify_track_id(&state.pool, video_id)
-        .await
-        .unwrap();
-    assert_eq!(stored.as_deref(), Some("3n3Ppam7vgaVa1iaRUc9Lp"));
-}
-
-#[tokio::test]
-async fn patch_video_empty_spotify_url_clears_track_id() {
-    let state = test_state().await;
-    let video_id = insert_test_video(&state).await;
-    // Pre-set a value.
-    db::models::set_video_spotify_track_id(&state.pool, video_id, Some("3n3Ppam7vgaVa1iaRUc9Lp"))
-        .await
-        .unwrap();
-    let app = app(state.clone());
-
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/v1/videos/{video_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(
-                    serde_json::to_string(&serde_json::json!({"spotify_url": ""})).unwrap(),
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NO_CONTENT);
-
-    let stored = db::models::get_video_spotify_track_id(&state.pool, video_id)
-        .await
-        .unwrap();
-    assert!(
-        stored.is_none(),
-        "empty spotify_url must clear track ID to NULL"
-    );
-}
-
-#[tokio::test]
-async fn patch_video_malformed_spotify_url_returns_400() {
-    let state = test_state().await;
-    let video_id = insert_test_video(&state).await;
-    let app = app(state);
-
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/v1/videos/{video_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(
-                    serde_json::to_string(&serde_json::json!({"spotify_url": "not a valid url"}))
-                        .unwrap(),
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
-
-#[tokio::test]
-async fn patch_video_empty_body_still_returns_400_after_spotify_field_added() {
-    let state = test_state().await;
-    let video_id = insert_test_video(&state).await;
-    let app = app(state);
-
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/v1/videos/{video_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(
-                    serde_json::to_string(&serde_json::json!({})).unwrap(),
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
